MOVED EVERTYHING IN APP.JS IN DIFFERENT COMPONENTS IDK WHY I DIDNT DO THAT FROM THE START 😤
